fix: compute router_logits in fp32 for better gating precision
Cast hidden_states to fp32 before gate linear to ensure the router
logits matmul happens in fp32 precision, improving expert selection
accuracy especially at topk boundaries.

+++ vllm_ascend/patch/worker/patch_deepseekv3.py
@@ -130,7 +130,7 @@ class AscendDeepseekV2MoE(DeepseekV2MoE, nn.Module):
         if self.is_sequence_parallel:
             hidden_states = sequence_parallel_chunk(hidden_states)
 
-        router_logits, _ = self.gate(hidden_states)
+        router_logits, _ = self.gate(hidden_states.float())
         fused_moe_out = self.experts(hidden_states=hidden_states,
                                      router_logits=router_logits)
         shared_output, final_hidden_states = fused_moe_out
